+++ src/Data/_Int64.cs
@@ -775,7 +775,7 @@ namespace DevZest.Data
             return base.GetHashCode();
         }
 
-        internal Identity SetIdentity(int seed, int increment)
+        internal Identity SetIdentity(long seed, long increment)
         {
             VerifyDesignMode();
